backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
faiss-cpu==1.7.4
google-generativeai==0.3.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
pypdf==3.17.1
numpy==1.24.3
pydantic==2.4.2
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4 